+++ pages/basePageComponents/footerPage.ts
