+++ pages/inventoryPageComponents/inventoryCard.ts
