+++ pages/productPage.ts
